mkd_note_taking_app/requirements.txt
Django==5.1.1
Markdown==3.7
pyspellchecker==0.8.1
